+++ modules/virt-manager.nix
@@ -1,0 +1,13 @@
+{
+  nixosModule = { user, ... }: {
+    virtualisation.libvirtd.enable = true;
+
+    users.users.${user}.extraGroups = [ "libvirtd" ];
+  };
+
+  hmModule = { pkgs, ... }: {
+    home.packages = builtins.attrValues {
+      inherit (pkgs) virt-manager;
+    };
+  };
+}